+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.35 - 2026-06-08
+### Fixed
+- Retry mechanism for failing Nitecrawler serial comm
+- Share invalid chars across platforms
+
 ## 1.1.34 - 2026-06-02
 ### Added
 - phd2 profile properties